// vendor/raw.githubusercontent.com/#mikaelporttila_0c368/rss/refs/tags/1.0.3/src/types/fields/atom_fields.ts
export enum AtomFields {
  Author = "author",
  Category = "category",
  Content = "content",
  Contributor = "contributor",
  Email = "email",
  Entry = "entry",
  Feed = "feed",
  Href = "href",
  Icon = "icon",
  Id = "id",
  Link = "link",
  Name = "name",
  Published = "published",
  Rel = "rel",
  Rights = "rights",
  Source = "source",
  Src = "src",
  Summary = "summary",
  Title = "title",
  Type = "type",
  Updated = "updated",
  Uri = "uri",
  Value = "value",
  FeedburnerOrigLink = "feedburner:origlink",
}